+++ components/datetime/tests/patterns/tests/timezones.json
@@ -4,8 +4,7 @@
     "config": {
       "time_zone_id": "America/Los_Angeles",
       "metazone_id": "America_Pacific",
-      "time_variant": "daylight",
-      "country_code": null
+      "time_variant": "daylight"
     },
     "date_time": "2021-01-11T12:00:00.000-07:00",
     "expectations": [
@@ -101,8 +100,7 @@
     "config": {
       "time_zone_id": "Asia/Tokyo",
       "metazone_id": "Japan",
-      "time_variant": null,
-      "country_code": null
+      "time_variant": null
     },
     "date_time": "2021-01-11T12:00:00.000+09:00",
     "expectations": [
@@ -198,8 +196,7 @@
     "config": {
       "time_zone_id": "Antarctica/Troll",
       "metazone_id": "GMT",
-      "time_variant": null,
-      "country_code": null
+      "time_variant": null
     },
     "date_time": "2021-01-11T12:00:00.000Z",
     "expectations": [
@@ -310,8 +307,7 @@
     "config": {
       "time_zone_id": "Asia/Katmandu",
       "metazone_id": "Nepal",
-      "time_variant": null,
-      "country_code": null
+      "time_variant": null
     },
     "date_time": "2021-01-11T12:00:00.000+05:45",
     "expectations": [
